script modifies /etc/fstab
URL: https://raw.githubusercontent.com/MicrosoftDocs/mslearn-add-and-size-disks-in-azure-virtual-machines/master/add-data-disk.sh

The script modifies /etc/fstab. Improperly modifying the /etc/fstab file could result in an unbootable system. Always test configuration changes on a temporary scratch system before you deploy to production. Refer to your distribution's documentation to learn how to properly modify this file. In production, we also recommend that you create a backup of this file so you can restore the configuration if needed.


To summarize, the script:
Partitions the drive /dev/sdc.
Creates an ext4 filesystem on the drive.
Create the /uploads directory, which we use as our mount point.
Attaches the disk to the mount point.
Updates /etc/fstab so that the drive is mounted automatically after the system reboots.